Split out player selector UI specific to game setup
