+++ src/main/kotlin/dartzee/screen/GameSetupScreen.kt
@@ -27,7 +27,7 @@ class GameSetupScreen : EmbeddedScreen()
     private val panelPlayers = JPanel()
     private val launchPanel = JPanel()
     val btnLaunch = JButton("Launch Game")
-    val playerSelector = PlayerSelector()
+    val playerSelector = GameSetupPlayerSelector()
     val gameTypeComboBox = ComboBoxGameType()
     private val panelGameTypeCb = JPanel()
     var gameParamFilterPanel: GameParamFilterPanel = GameParamFilterPanelX01()
